# Refactor internals to support realm concept