--- a/Jint.Repl/Program.cs
+++ b/Jint.Repl/Program.cs
@@ -67,7 +67,7 @@ namespace Jint.Repl
                     var result = engine.Evaluate(input, parserOptions);
                     if (!result.IsNull() && !result.IsUndefined())
                     {
-                        var str = TypeConverter.ToString(engine.Json.Stringify(engine.Json, Arguments.From(result, Undefined.Instance, "  ")));
+                        var str = TypeConverter.ToString(engine.Realm.Intrinsics.Json.Stringify(engine.Realm.Intrinsics.Json, Arguments.From(result, Undefined.Instance, "  ")));
                         Console.WriteLine(str);
                     }
                     else