--- a/Jint/Host/DefaultHost.cs
+++ b/Jint/Host/DefaultHost.cs
@@ -1,0 +1,9 @@
+namespace Jint
+{
+    internal class DefaultHost : Host
+    {
+        public DefaultHost(Engine engine) : base(engine)
+        {
+        }
+    }
+}