--- a/Jint/Host/IRealmFactory.cs
+++ b/Jint/Host/IRealmFactory.cs
@@ -1,0 +1,7 @@
+namespace Jint
+{
+    public interface IRealmFactory
+    {
+        public Realm CreateRealm(Engine engine);
+    }
+}