This PR will hopefully produce first version of [Realm](https://tc39.es/ecma262/#sec-code-realms) and [Host](https://tc39.es/ecma262/#sec-hosts-and-implementations) concepts.

 This will cater way for to additional features:

* [Isolated execution context,](https://github.com/sebastienros/jint/pull/789) we can have custom global built and some callbacks
* BrowserHost (or something similar) can be implemented for scenarios like [AngleSharp.JS](https://github.com/AngleSharp/AngleSharp.Js/issues/66), where default environment is not what is wanted
* Having separate initialization could also mean that we could reset the global environment without ditching the entire engine

This PR will also unify and simplify initial engine by introducing separate `Intrinsics` object instead having it all in `Engine`.