@sebastienros 2021-06-11
What's your feeling about how much closer it is to the specs with this PR?