@lahma 2021-06-12
Current status is that it should now follow spec and constructs quite closely. I *think* that the constructs are now easier to understand and one can follow the spec with regards to references to realms (current realm etc). Tests are passing and now it's time for some polish and checking if AngleSharp custom host (Window object) can be implemented and what kind of hosting API it requires, close to finish I would say.