@lahma 2021-06-15
Because this is already quite massive change I think it would be nice to get this into main before continuing the integration story via Host concepts. I've tested this API-wise against RavenDB and AngleSharp.JS, engine.Obsolete.cs helps to transition to realm and their code compiles as usual.