@lahma 2021-06-15
Another nice outcome of making engine initialization lazier is that now one gets to run minimal scripts faster:

```csharp
var engine = new Engine();
return engine.Evaluate("return [].length + ''.length").AsNumber();
```

*Before*

|      Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------ |---------:|----------:|----------:|-------:|-------:|------:|----------:|
| BuildEngine | 5.587 us | 0.0265 us | 0.0207 us | 1.2894 | 0.0687 |     - |  21.09 KB |

*After*

|      Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------ |---------:|----------:|----------:|-------:|-------:|------:|----------:|
| BuildEngine | 4.312 us | 0.0159 us | 0.0149 us | 0.8545 | 0.0229 |     - |  14.04 KB |

The change is more visible when the script is parsed beforehand and only the Script instance is executed against new engine:

*Before*

|      Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------ |---------:|----------:|----------:|-------:|-------:|------:|----------:|
| BuildEngine | 3.323 us | 0.0163 us | 0.0136 us | 1.0071 | 0.0381 |     - |  16.48 KB |

*After*

|      Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------ |---------:|----------:|----------:|-------:|-------:|------:|----------:|
| BuildEngine | 2.366 us | 0.0109 us | 0.0102 us | 0.5760 | 0.0153 |     - |   9.43 KB |