@lahma 2021-06-15
[Example how spec can now be followed using the new constructs](https://github.com/sebastienros/jint/pull/907/files#diff-1b5cc28bf240558bbe821785127badf825812d58b96abfa497ea270066523cf1L426).